Make tests/lora/utils usable by plugins
